+++ View/Education/education.php
@@ -31,7 +31,7 @@
             <div class="collapse navbar-collapse " id="navbarNavAltMarkup">
                 <div class="navbar-nav ml-auto">
                     <a class="nav-item nav-link active text-dark font-weight-bold" href="../index.php">Home</a>
-                    <a class="nav-item nav-link text-dark" href="../Salary/salaryAdmin.php">Salary</a>
+                    <a class="nav-item nav-link text-dark" href="../Salary/salary.php">Salary</a>
                     <a class="nav-item nav-link text-dark" href="../Employee/employee.php">Employee</a>
                     <a class="nav-item nav-link text-dark" href="../Department/department.php">Department</a>
                     <a class="nav-item nav-link text-dark" href="../Job/job.php">Job Roles</a>
@@ -48,7 +48,9 @@
         <div class="row">
             <div class="col-md-2 col-sm-12">
                 <a href="../index.php" class="btn btn-primary">Back</a>
-                <a href="./addEducation.php" class="btn btn-warning my-3">Add Education Option</a>
+                <?php if ($_SESSION['role'] === 'admin') : ?>
+                    <a href="./addEducation.php" class="btn btn-warning my-3">Add Education Option</a>
+                <?php endif; ?>
             </div>
             <div class="col-md-10 col-sm-12 table-responsive">
                 <p class="text-danger font-weight-bold my-0">*slide right to see more (for smaller screens)</p>
@@ -58,7 +60,9 @@
                             <th scope="col">#</th>
                             <th scope="col">Education Term Name</th>
                             <th scope="col">Number of employees</th>
-                            <th scope="col">Actions</th>
+                            <?php if ($_SESSION['role'] === 'admin') : ?>
+                                <th scope="col">Actions</th>
+                            <?php endif; ?>
                         </tr>
                     </thead>
                     <tbody>
@@ -75,10 +79,12 @@
                                 <th scope="row"><?= $number++ ?></th>
                                 <td><?= $fetchEducationData['education_term'] ?></td>
                                 <td><?= $fetchEducationData['matched_employee_education'] ?></td>
-                                <td>
-                                    <a href="editEducation.php?editinfo=<?= $fetchEducationData['educationId'] ?>" class="btn btn-success edit-btn">Edit</a>
-                                    <a href="../../Model/educationController.php?deleteEducation=<?= $fetchEducationData['educationId'] ?>" class="btn btn-danger" onclick="return confirm('Do you want to delete this employee record ?')">Delete</a>
-                                </td>
+                                <?php if ($_SESSION['role'] === 'admin') : ?>
+                                    <td>
+                                        <a href="editEducation.php?editinfo=<?= $fetchEducationData['educationId'] ?>" class="btn btn-success edit-btn">Edit</a>
+                                        <a href="../../Model/educationController.php?deleteEducation=<?= $fetchEducationData['educationId'] ?>" class="btn btn-danger" onclick="return confirm('Do you want to delete this employee record ?')">Delete</a>
+                                    </td>
+                                <?php endif; ?>
                             </tr>
                         <?php } ?>
                     </tbody>
